extract seed 39 components
