+++ pages/seed/39/index.tsx
@@ -1,4 +1,4 @@
-import { Card, CardContent, Grid, Typography, useMediaQuery } from '@mui/material';
+import { Card, CardContent, useMediaQuery } from '@mui/material';
 import { useTheme } from '@mui/system';
 import { useCallback } from 'react';
 import useWindowDimensions from '@hooks/useWindowDimensions';
@@ -11,43 +11,11 @@ import { Comments } from '@components/comments';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { useTranslation } from 'next-i18next';
 import useI18nSeoProps from '@components/seo/useI18nSeoProps';
+import { QuestionAnswerItem } from '@components/seed/39';
 
 interface Seed39Props {
     data: QuestionAnswer[];
 }
-
-interface QuestionAnswerItemProps extends QuestionAnswer {
-}
-
-const QuestionAnswerItem = (props: QuestionAnswerItemProps) => {
-    const { question, answer, choices } = props;
-    return (
-        <Grid container alignItems='center' spacing={1}>
-            <Grid item xs={12} sm={12} md={12} lg={4}>
-                <Typography variant={'body1'}>{question}</Typography>
-            </Grid>
-            {
-                choices.map((choice, i) => (
-                    <Grid item xs={12} sm={12} md={3} lg={2} key={`${choice}-${answer}`}>
-                        <Card sx={theme => ({
-                            bgcolor: answer === i
-                                ? theme.palette.primary.light
-                                : theme.palette.background.default,
-                            color: answer === i
-                                ? theme.palette.getContrastText(theme.palette.primary.light)
-                                : theme.palette.getContrastText(theme.palette.background.default),
-                        })}
-                              elevation={0}>
-                            <CardContent sx={(theme) => ({ padding: theme.spacing(1.5) })}>
-                                <Typography variant={'subtitle2'}>{i + 1}. {choice}</Typography>
-                            </CardContent>
-                        </Card>
-                    </Grid>
-                ))
-            }
-        </Grid>
-    );
-};
 
 const Seed39 = (props: Seed39Props) => {
     const { t, i18n } = useTranslation();
