+++ src/components/seed/39/index.ts
@@ -1,0 +1,1 @@
+export { default as QuestionAnswerItem } from './QuestionAnswerItem';
